More REST improvments and lots of testing with wake triggers.
REST tidied up a bit. I might start working on creating the REST client
soon, so the ESO8266 can fire an event to an external REST server.

I have been trying to get the ESP8266 to be stable so it can run for
ling periods of time.
After spending hours trying to get an external trigger to work properly,
the circuit got too complex and the power consumption went too high with
an external trigger source.

So i went back and did a bunch of testing with the old capacitor on
GPIO14 and i think its stable now.
The issue was the sudden current draw on the pin was causing the IC to
be unstable.
I tried a 1K, 4.7K and 10K resistor between the capacitor and GPIO14.
I wanted to use a low as possible resistor so it doesn't take too long
to charge, but this turned out the be the oppsite of what was best.
The best option was for GPIO14 to have a 10K resistor between it and the
capacitor, and just set the charge time to 100ms, which seems high, but
the ESP8266 seems to be in sleep a lot more than with lower times, and
stable. so its a win-win for unknown reasons.

So take my word.... Use a task in user_main to handle the timing, Dont
bother with an external trigger, its just more hassle and power than its
worth.

All my testing is now done with a server making a REST request every 60
seconds.

Ill try draw a scematic soon.