First succesful compile with workign REST client.
WARNING, client is not stable!
Looks like a memory leak or sharing issue. Ill fix when i get time
again.
The code layout is still not quite how i wan it.. I want the REST client
to interface with the enpoint in a simular way as the REST server.

I will be changing the way the ADC is controlled. It will only have min
and max and be up to the hub sever to return new values to prevent it
always getting triggeres.

Still need to make the REST client set endpoint values using the result
body so the request are mirrored with the ones sent to the ESP.

Thought i better save my work!

If you use this release, disable RESTEndpointADCCheckStatus()